rag/requirements.txt
pathway
python-dotenv
google-generativeai
